fixed conftests, fixed .coverage gitignore, and removed extra deletes from login tests
