+++ Src/banking_service/banking_tests/test_banking.py
@@ -1,30 +1,6 @@
 """Test File for Banking Service"""
 
 import pytest
-from fastapi.testclient import TestClient
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker
-from Src.banking_service.banking import app, get_db
-from Src.banking_service.models import Base
-
-TEST_DB_URL = "sqlite+pysqlite:///:memory:"
-engine = create_engine(TEST_DB_URL, connect_args={"check_same_thread": False})
-TestingSessionLocal = sessionmaker(bind=engine, expire_on_commit=False)
-Base.metadata.create_all(bind=engine)
-
-@pytest.fixture
-def client():
-    def override_get_db():
-        db = TestingSessionLocal()
-        try:
-            yield db
-        finally:
-            db.close()
-            app.dependency_overrides[get_db] = override_get_db
-    with TestClient(app) as c:
-        # hand the client to the test
-        yield c
-        # --- teardown happens when the 'with' block exits ---
 
 def bank_account_details_payload(id=1, name="John", email="John@Example.com", pin=1225, card="1000000000000000", balance=2):
     return {"id": id, "name": name, "email": email, "pin": pin, "card": card, "balance": balance}
